refactor(auth): restructure source code
